+++ yospace/src/main/java/com/bitmovin/player/integration/yospace/YospaceLiveInitializationType.kt
@@ -1,6 +1,6 @@
 package com.bitmovin.player.integration.yospace
 
-enum class YospaceLiveInitialisationType {
+enum class YospaceLiveInitializationType {
     PROXY,
     DIRECT
 }