+++ yospace/src/main/java/com/bitmovin/player/integration/yospace/config/YospaceConfig.kt
@@ -1,15 +1,14 @@
 package com.bitmovin.player.integration.yospace.config
 
-import com.bitmovin.player.integration.yospace.YospaceLiveInitialisationType
+import com.bitmovin.player.integration.yospace.YospaceLiveInitializationType
 
 data class YospaceConfig(
     val userAgent: String? = null,
-    val readTimeout: Int = 25_000,
-    val connectTimeout: Int = 25_000,
     val requestTimeout: Int = 25_000,
-    val liveInitialisationType: YospaceLiveInitialisationType = YospaceLiveInitialisationType.DIRECT,
+    val liveInitializationType: YospaceLiveInitializationType = YospaceLiveInitializationType.DIRECT,
     val isDebug: Boolean = false,
-    val filterMetadataType: MetadataType? = MetadataType.EMSG
+    val filterMetadataType: MetadataType? = MetadataType.EMSG,
+    val yospaceDebugMode: YospaceDebugMode = YospaceDebugMode.NONE
 )
 
 public enum class MetadataType{