+++ yospace/src/main/java/com/bitmovin/player/integration/yospace/config/YospaceDebugMode.kt
@@ -1,0 +1,7 @@
+package com.bitmovin.player.integration.yospace.config
+
+enum class YospaceDebugMode {
+    NONE,
+    VALIDATION,
+    ALL
+}